feat: rename Number to NumberUtil
