--- src/main/java/de/jonahd345/extendedeconomy/util/Number.java
@@ -1,34 +1,0 @@
-package de.jonahd345.extendedeconomy.util;
-
-import java.text.DecimalFormat;
-
-public class Number {
-    private DecimalFormat format;
-
-    public Number() {
-        //this.format = new DecimalFormat("###,###,###,###,###,###,###,###,###");
-        this.format = new DecimalFormat("#,##0.00");
-    }
-
-    public String formatNumber(Object number) {
-        return this.format.format(number); //.replace(",", ".");
-    }
-
-    public boolean isDouble(String argument) {
-        try {
-            Double.parseDouble(argument);
-            return true;
-        } catch (Exception e) {
-            return false;
-        }
-    }
-
-    public boolean isInt(String argument) {
-        try {
-            Integer.parseInt(argument);
-            return true;
-        } catch (Exception e) {
-            return false;
-        }
-    }
-}
